Added ocamlbuild system w/ scripted C file gen
- bv{8,16,32,64}.[ch] generated by myocamlbuild.ml.
- can link and build entire library. God I hope this scales.

- I had issues with bv* in the bitvector subdirectory, even with
  <bitvector>:include. I'm really not sure why things did not parse well.

- I've left the build files in the root directory. This may be annoying in the
  future having to add lib/ to everything, but I would like to keep the build
  files and the source files separate.

+++ lib/nonAdditive_o.ml
@@ -2,7 +2,7 @@ type t =
   { codes : int array array;
     weights : float array;
     states : int array;
-    cost : int; }
+    cost : float; }
 
 type spec = t
